Add support for React Native Web
This is part refactor, part rebuild.

In order to add support to RN web I created a new instance of CRNA with the `--with-web-support` flag set.

I then removed the gifted chat package, with the sole purpose of getting this to me the simplest implementation possible, the package can be re-added at a later stage.